+++ Exec/DoubleRamp/GNUmakefile
@@ -1,0 +1,39 @@
+CAMR_HOME      ?= ../..
+AMREX_HOME     ?= ../../../amrex
+
+# AMReX
+DIM = 2
+COMP = gnu
+PRECISION = DOUBLE
+
+# Profiling
+PROFILE = FALSE
+#TINY_PROFILE = TRUE
+COMM_PROFILE = FALSE
+TRACE_PROFILE = FALSE
+MEM_PROFILE = FALSE
+USE_GPROF = FALSE
+
+# Performance
+USE_MPI = TRUE
+USE_OMP = FALSE
+USE_CUDA = FALSE
+USE_HIP = FALSE
+USE_SYCL = FALSE
+
+# Debugging
+DEBUG = FALSE
+DEBUG = TRUE
+
+TEST = TRUE
+USE_ASSERTION = TRUE
+
+# CAMR
+USE_EB = TRUE
+Eos_Model := GammaLaw
+
+# GNU Make
+Bpack := ./Make.package
+Blocs := .
+
+include $(CAMR_HOME)/Exec/Make.CAMR
