+++ Exec/DoubleRamp/Make.package
@@ -1,0 +1,3 @@
+CEXE_headers += prob.H
+CEXE_headers += prob_parm.H
+CEXE_sources += prob.cpp
