+++ Source/Hydro/CAMR_hydro.H
@@ -1,5 +1,5 @@
-#ifndef HYDRO_H
-#define HYDRO_H
+#ifndef CAMR_HYDRO_H
+#define CAMR_HYDRO_H
 
 #include "CAMR.H"
 #include "IndexDefines.H"
@@ -8,222 +8,7 @@
 #include <AMReX_EBFluxRegister.H>
 #endif
 
-AMREX_GPU_DEVICE
-AMREX_FORCE_INLINE
-void
-CAMR_divu(
-  const int i,
-  const int j,
-  const int k,
-  amrex::Array4<const amrex::Real> const& q,
-  AMREX_D_DECL(const amrex::Real dx,
-               const amrex::Real dy,
-               const amrex::Real dz),
-  amrex::Array4<amrex::Real> const& divu,
-  const int* domlo, const int* domhi,
-  const int* bclo , const int* bchi)
-{
-    int bclo_x = bclo[0]; int domlo_x = domlo[0];
-    int bclo_y = bclo[1]; int domlo_y = domlo[1];
-    int bchi_x = bchi[0]; int domhi_x = domhi[0];
-    int bchi_y = bchi[1]; int domhi_y = domhi[1];
-
-#if (AMREX_SPACEDIM == 2)
-    amrex::Real ux, vy;
-
-    if (j == domlo_y && bclo_y == NoSlipWall) {
-        ux = ( q(i,j  ,k,QU) - q(i-1,j  ,k,QU) ) / dx;
-    } else if (j == domhi_y+1 && bchi_y == NoSlipWall) {
-        ux = (q(i,j-1,k,QU) - q(i-1,j-1,k,QU)) / dx;
-    } else {
-        ux = 0.5 * (q(i,j  ,k,QU) - q(i-1,j  ,k,QU)
-                 +  q(i,j-1,k,QU) - q(i-1,j-1,k,QU)) / dx;
-    }
-
-    if (i == domlo_x && bclo_x == NoSlipWall) {
-      vy = ( q(i  ,j,k,QV) - q(i  ,j-1,k,QV) ) / dy;
-    } else if (i == domhi_x+1 && bchi_x == NoSlipWall) {
-      vy = ( q(i-1,j,k,QV) - q(i-1,j-1,k,QV) ) / dy;
-    } else {
-      vy = 0.5 * ( q(i  , j, k, QV) - q(i  ,j-1,k,QV) +
-                   q(i-1, j, k, QV) - q(i-1,j-1,k,QV) ) / dy;
-    }
-
-#elif (AMREX_SPACEDIM == 3)
-    amrex::Real ux, vy, wz;
-
-    int bclo_z = bclo[2]; int domlo_z = domlo[2];
-    int bchi_z = bchi[2]; int domhi_z = domhi[2];
-
-    ux = 0.25 * (
-        q(i, j  , k  , QU) - q(i-1, j  , k  , QU) +
-        q(i, j  , k-1, QU) - q(i-1, j  , k-1, QU) +
-        q(i, j-1, k  , QU) - q(i-1, j-1, k  , QU) +
-        q(i, j-1, k-1, QU) - q(i-1, j-1, k-1, QU)) / dx;
-    vy = 0.25 * (
-        q(i  , j, k  , QV) - q(i  , j-1, k  , QV) +
-        q(i  , j, k-1, QV) - q(i  , j-1, k-1, QV) +
-        q(i-1, j, k  , QV) - q(i-1, j-1, k  , QV) +
-        q(i-1, j, k-1, QV) - q(i-1, j-1, k-1, QV)) / dy;
-    wz = 0.25 * (
-        q(i  , j  , k, QW) - q(i  , j  , k-1, QW) +
-        q(i-1, j  , k, QW) - q(i-1, j  , k-1, QW) +
-        q(i  , j-1, k, QW) - q(i  , j-1, k-1, QW) +
-        q(i-1, j-1, k, QW) - q(i-1, j-1, k-1, QW)) / dz;
-
-    if (i == domlo_x && bclo_x == NoSlipWall) {
-        vy = 0.5 * (
-            q(i  , j, k  , QV) - q(i  , j-1, k  , QV) +
-            q(i  , j, k-1, QV) - q(i  , j-1, k-1, QV) ) / dy;
-        wz = 0.5 * (
-            q(i  , j  , k, QW) - q(i  , j  , k-1, QW) +
-            q(i  , j-1, k, QW) - q(i  , j-1, k-1, QW) ) / dz;
-    } else if (i == domhi_x+1 && bchi_x == NoSlipWall) {
-        vy = 0.5 * (
-            q(i-1, j, k  , QV) - q(i-1, j-1, k  , QV) +
-            q(i-1, j, k-1, QV) - q(i-1, j-1, k-1, QV) ) / dy;
-        wz = 0.5 * (
-            q(i-1, j  , k, QW) - q(i-1, j  , k-1, QW) +
-            q(i-1, j-1, k, QW) - q(i-1, j-1, k-1, QW) ) / dz;
-    }
-
-    if (j == domlo_y && bclo_y == NoSlipWall) {
-        ux = 0.5 * (
-            q(i, j  , k  , QU) - q(i-1, j  , k  , QU) +
-            q(i, j  , k-1, QU) - q(i-1, j  , k-1, QU) ) / dx;
-        wz = 0.5 * (
-            q(i  , j  , k, QW) - q(i  , j  , k-1, QW) +
-            q(i-1, j  , k, QW) - q(i-1, j  , k-1, QW) ) / dz;
-    } else if (j == domhi_y+1 && bchi_y == NoSlipWall) {
-        ux = 0.5 * (
-            q(i, j-1, k  , QU) - q(i-1, j-1, k  , QU) +
-            q(i, j-1, k-1, QU) - q(i-1, j-1, k-1, QU) ) / dx;
-        wz = 0.5 * (
-            q(i  , j-1, k, QW) - q(i  , j-1, k-1, QW) +
-            q(i-1, j-1, k, QW) - q(i-1, j-1, k-1, QW)) / dz;
-    }
-
-    if (k == domlo_z && bclo_z == NoSlipWall) {
-        ux = 0.5 * (
-            q(i, j  , k  , QU) - q(i-1, j  , k  , QU) +
-            q(i, j-1, k  , QU) - q(i-1, j-1, k  , QU) ) / dx;
-        vy = 0.5 * (
-            q(i  , j, k  , QV) - q(i  , j-1, k  , QV) +
-            q(i-1, j, k  , QV) - q(i-1, j-1, k  , QV) ) / dy;
-    } else if (k == domhi_z+1 && bchi_z == NoSlipWall) {
-        ux = 0.5 * (
-            q(i, j  , k-1, QU) - q(i-1, j  , k-1, QU) +
-            q(i, j-1, k-1, QU) - q(i-1, j-1, k-1, QU) ) / dx;
-        vy = 0.5 * (
-            q(i  , j, k-1, QV) - q(i  , j-1, k-1, QV) +
-            q(i-1, j, k-1, QV) - q(i-1, j-1, k-1, QV)) / dy;
-    }
-
-    // Fix ux at y-z corners
-    if (j == domlo_y && bclo_y == NoSlipWall &&
-        k == domlo_z && bclo_z == NoSlipWall) {
-        ux = ( q(i  , j  , k, QU) - q(i-1, j  , k  , QU) ) / dx;
-    } else if ( (j == domhi_y+1 && bchi_y == NoSlipWall) &&
-                (k == domlo_z   && bclo_z == NoSlipWall) ) {
-        ux = ( q(i  , j-1, k, QU) - q(i-1, j-1, k  , QU) ) / dx;
-    } else if ( (j == domlo_y   && bclo_y == NoSlipWall) &&
-                (k == domhi_z+1 && bchi_z == NoSlipWall) ) {
-        ux = ( q(i  , j  , k-1, QU) - q(i-1, j  , k-1, QU) )/ dx;
-    } else if ( (j == domhi_y+1 && bchi_y == NoSlipWall) &&
-                (k == domhi_z+1 && bchi_z == NoSlipWall) ) {
-        ux = ( q(i  , j-1, k-1, QU) - q(i-1, j-1, k-1, QU) )/ dx;
-    }
-
-    // Fix vy at x-z corners
-    if (i == domlo_x && bclo_x == NoSlipWall &&
-        k == domlo_z && bclo_z == NoSlipWall) {
-        vy = ( q(i  , j  , k, QV) - q(i  , j-1, k  , QV) ) / dy;
-    } else if ( (i == domhi_x+1 && bchi_x == NoSlipWall) &&
-                (k == domlo_z   && bclo_z == NoSlipWall) ) {
-        vy = ( q(i-1, j  , k, QV) - q(i-1, j-1, k  , QV) ) / dy;
-    } else if ( (i == domlo_x   && bclo_x == NoSlipWall) &&
-                (k == domhi_z+1 && bchi_z == NoSlipWall) ) {
-        vy = ( q(i  , j  , k-1, QV) - q(i  , j-1, k-1, QV) )/ dy;
-    } else if ( (i == domhi_x+1 && bchi_x == NoSlipWall) &&
-                (k == domhi_z+1 && bchi_z == NoSlipWall) ) {
-        vy = ( q(i-1, j  , k-1, QV) - q(i-1, j-1, k-1, QV) )/ dy;
-    }
-
-    // Fix wz at x-y corners
-    if (i == domlo_x && bclo_x == NoSlipWall &&
-        j == domlo_y && bclo_y == NoSlipWall) {
-        wz = ( q(i  , j  , k, QW) - q(i  , j  , k-1, QW) ) / dz;
-    } else if ( (i == domhi_x+1 && bchi_x == NoSlipWall) &&
-                (j == domlo_y   && bclo_y == NoSlipWall) ) {
-        wz = ( q(i-1, j  , k, QW) - q(i-1, j  , k-1, QW) ) / dz;
-    } else if ( (i == domlo_x   && bclo_x == NoSlipWall) &&
-                (j == domhi_y+1 && bchi_y == NoSlipWall) ) {
-        wz = ( q(i  , j-1, k, QW) - q(i  , j-1, k-1, QW) )/ dz;
-    } else if ( (i == domhi_x+1 && bchi_x == NoSlipWall) &&
-                (j == domhi_y+1 && bchi_y == NoSlipWall) ) {
-        wz = ( q(i-1, j-1, k, QW) - q(i-1, j-1, k-1, QW) )/ dz;
-    }
-#endif
-
-  divu(i, j, k) = AMREX_D_TERM(ux, +vy, +wz);
-}
-
-AMREX_GPU_DEVICE
-AMREX_FORCE_INLINE
-void
-CAMR_norm_spec_flx(
-  const int i, const int j, const int k, amrex::Array4<amrex::Real> const& flx)
-{
-  const int num = NUM_SPECIES + UFS;
-  amrex::Real summ, fac;
-  summ = 0.0;
-  for (int n = UFS; n < num; ++n)
-    summ += flx(i, j, k, n);
-
-  fac = (summ != 0.0) ? flx(i, j, k, URHO) / summ : 1.0;
-  for (int n = UFS; n < num; ++n)
-    flx(i, j, k, n) *= fac;
-}
-
-AMREX_GPU_DEVICE
-AMREX_FORCE_INLINE
-void
-CAMR_ext_flx(
-  const int i,
-  const int j,
-  const int k,
-  amrex::Array4<amrex::Real> const& flx,
-  amrex::Array4<const amrex::Real> const& a)
-{
-  amrex::Real area = a(i, j, k);
-  for (int n = 0; n < NVAR; ++n)
-    flx(i, j, k, n) *= area;
-}
-
-AMREX_GPU_DEVICE
-AMREX_FORCE_INLINE
-void
-CAMR_update(
-  const int i,
-  const int j,
-  const int k,
-  amrex::Array4<amrex::Real> const& update,
-  const amrex::GpuArray<const amrex::Array4<amrex::Real>, AMREX_SPACEDIM> flx,
-  amrex::Array4<const amrex::Real> const& vol,
-  amrex::Array4<const amrex::Real> const& pdivu)
-{
-    amrex::Real volinv = 1.0/vol(i, j, k);
-    for (int n = 0; n < NVAR; ++n) {
-      update(i, j, k, n) +=
-        (AMREX_D_TERM(flx[0](i, j, k, n) - flx[0](i + 1, j, k, n),
-                      +flx[1](i, j, k, n) - flx[1](i, j + 1, k, n),
-                      +flx[2](i, j, k, n) - flx[2](i, j, k + 1, n))) *
-        volinv;
-    }
-    update(i, j, k, UEINT) -= pdivu(i, j, k);
-}
-
-void CAMR_umdrv( bool do_mol,
+void CAMR_umdrv (bool do_mol,
                  amrex::Box const& bx,
                  amrex::Geometry const& geom,
                  const int* bclo, const int* bchi,
@@ -247,7 +32,7 @@ void CAMR_umdrv( bool do_mol,
                  const amrex::GpuArray<const amrex::Array4<const amrex::Real>, AMREX_SPACEDIM> a,
                  amrex::Array4<amrex::Real> const& vol);
 
-void adjust_fluxes(
+void adjust_fluxes (
   amrex::Box const& bx,
   amrex::Array4<const amrex::Real> const& u_arr,
   const amrex::GpuArray<const amrex::Array4<amrex::Real>, AMREX_SPACEDIM> flx,
@@ -257,7 +42,7 @@ void adjust_fluxes(
   const int* domlo, const int* domhi, const int* bclo, const int* bchi,
   amrex::Real const difmag);
 
-void CAMR_consup(
+void CAMR_consup (
   amrex::Box const& bx,
   amrex::Array4<amrex::Real> const& update,
   const amrex::GpuArray<const amrex::Array4<amrex::Real>, AMREX_SPACEDIM> flx,
@@ -266,7 +51,7 @@ void CAMR_consup(
 
 #ifdef AMREX_USE_EB
 void
-CAMR_umdrv_eb( const bool do_mol, amrex::Box const& bx,
+CAMR_umdrv_eb (const bool do_mol, amrex::Box const& bx,
                amrex::Box const& bxg_i, const amrex::MFIter& mfi,
                amrex::Geometry const& geom,
                const amrex::EBFArrayBoxFactory* ebfact,
@@ -301,8 +86,23 @@ CAMR_umdrv_eb( const bool do_mol, amrex::Box const& bx,
                const std::string& l_redistribution_type,
                const int l_eb_weights_type);
 
+void adjust_fluxes_eb (
+               const amrex::Box& bx,
+               amrex::Array4<const amrex::Real> const& q_arr,
+               amrex::Array4<const amrex::Real> const& u_arr,
+               AMREX_D_DECL(amrex::Array4<amrex::Real const> const& apx,
+                            amrex::Array4<amrex::Real const> const& apy,
+                            amrex::Array4<amrex::Real const> const& apz ),
+               amrex::Array4<const amrex::Real> const& vfrac,
+               const amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> dx,
+               const amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> dxInv,
+               const amrex::GpuArray<const amrex::Array4<amrex::Real>, AMREX_SPACEDIM> flux,
+               const int* domlo, const int* domhi,
+               const int*  bclo, const int*  bchi,
+               const amrex::Real l_difmag);
+
 void
-CAMR_consup_eb( const amrex::Box& bx,
+CAMR_consup_eb (const amrex::Box& bx,
                 amrex::Array4<const amrex::Real> const& q_arr,
                 amrex::Array4<const amrex::Real> const& qaux_arr,
                 amrex::Array4<      amrex::Real> const& divc_arr,
