+++ Source/Hydro/CAMR_utils_eb_K.H
@@ -1,7 +1,245 @@
-#include <AMReX_Config.H>
-#include <AMReX_REAL.H>
-#include "MOL_hydro_K.H"
-#include "MOL_hypflux_K.H"
+#ifndef CAMR_UTILS_EB_H
+#define CAMR_UTILS_EB_H
+
+#ifdef AMREX_USE_EB
+
+#include "IndexDefines.H"
+#include <AMReX_EBFluxRegister.H>
+#include <CAMR_hypflux_eb_K.H>
+
+AMREX_GPU_DEVICE AMREX_FORCE_INLINE
+void
+CAMR_ext_flx_eb(
+  const int i,
+  const int j,
+  const int k,
+  amrex::Array4<amrex::Real> const& flx,
+  amrex::Real area,
+  amrex::Array4<const amrex::Real> const& aper)
+{
+  // The fluxes are multiplied by area fractions inside amrex
+  //  so here we just weight them by the regular areas
+  if (aper(i,j,k) > 0.) {
+    for (int n = 0; n < NVAR; ++n) {
+       flx(i, j, k, n) *= area;
+    }
+  }
+}
+
+AMREX_GPU_DEVICE AMREX_FORCE_INLINE
+void
+eb_divu (int i, int j, int k,
+         amrex::Array4<amrex::Real const> const& q,
+          amrex::Array4<amrex::Real> const& divu,
+          amrex::Array4<amrex::Real const> const& vfrac,
+          amrex::GpuArray<amrex::Real,AMREX_SPACEDIM> const& dxInv)
+{
+#if (AMREX_SPACEDIM == 2)
+    amrex::Real surrounding_vol = vfrac(i,j  ,k  ) + vfrac(i-1,j  ,k  ) +
+                                  vfrac(i,j-1,k  ) + vfrac(i-1,j-1,k  );
+    amrex::Real cutoff = 4.0 * (1.0 - 1.e-12);
+#elif (AMREX_SPACEDIM == 3)
+    amrex::Real surrounding_vol = vfrac(i,j  ,k  ) + vfrac(i-1,j  ,k  ) +
+                                  vfrac(i,j-1,k  ) + vfrac(i-1,j-1,k  ) +
+                                  vfrac(i,j  ,k-1) + vfrac(i-1,j  ,k-1) +
+                                  vfrac(i,j-1,k-1) + vfrac(i-1,j-1,k-1);
+    amrex::Real cutoff = 8.0 * (1.0 - 1.e-12);
+#endif
+
+    divu(i, j, k) = 0.0;
+
+    if (surrounding_vol >= cutoff) {
+
+#if (AMREX_SPACEDIM == 2)
+        amrex::Real ux = 0.5 *
+                         (q(i, j, k, QU) - q(i - 1, j, k, QU) + q(i, j - 1, k, QU) -
+                          q(i - 1, j - 1, k, QU)) * dxInv[0];
+        amrex::Real vy = 0.5 *
+                         (q(i, j, k, QV) - q(i, j - 1, k, QV) + q(i - 1, j, k, QV) -
+                          q(i - 1, j - 1, k, QV)) * dxInv[1];
+#elif (AMREX_SPACEDIM == 3)
+        amrex::Real ux = 0.25 *
+           (q(i, j, k, QU) - q(i - 1, j, k, QU) + q(i, j, k - 1, QU) -
+           q(i - 1, j, k - 1, QU) + q(i, j - 1, k, QU) - q(i - 1, j - 1, k, QU) +
+           q(i, j - 1, k - 1, QU) - q(i - 1, j - 1, k - 1, QU)) * dxInv[0];
+
+        amrex::Real vy = 0.25 *
+          (q(i, j, k, QV) - q(i, j - 1, k, QV) + q(i, j, k - 1, QV) -
+           q(i, j - 1, k - 1, QV) + q(i - 1, j, k, QV) - q(i - 1, j - 1, k, QV) +
+           q(i - 1, j, k - 1, QV) - q(i - 1, j - 1, k - 1, QV)) * dxInv[1];
+
+        amrex::Real wz = 0.25 *
+          (q(i, j, k, QW) - q(i, j, k - 1, QW) + q(i - 1, j, k, QW) -
+           q(i - 1, j, k - 1, QW) + q(i, j - 1, k, QW) - q(i, j - 1, k - 1, QW) +
+           q(i - 1, j - 1, k, QW) - q(i - 1, j - 1, k - 1, QW)) * dxInv[2];
+#endif
+      divu(i, j, k) = AMREX_D_TERM(ux, +vy, +wz);
+    }
+}
+
+AMREX_GPU_DEVICE AMREX_FORCE_INLINE
+void
+eb_pdivu (int i, int j, int k,
+          amrex::Array4<amrex::Real const> const& q,
+          AMREX_D_DECL(amrex::Array4<amrex::Real const> const& q1,
+                       amrex::Array4<amrex::Real const> const& q2,
+                       amrex::Array4<amrex::Real const> const& q3),
+          amrex::Array4<amrex::Real> const& divu,
+          amrex::Array4<amrex::EBCellFlag const> const& flag,
+          amrex::Array4<amrex::Real const> const& vfrc,
+          AMREX_D_DECL(amrex::Array4<amrex::Real const> const& apx,
+                       amrex::Array4<amrex::Real const> const& apy,
+                       amrex::Array4<amrex::Real const> const& apz),
+          AMREX_D_DECL(amrex::Array4<amrex::Real const> const& fcx,
+                       amrex::Array4<amrex::Real const> const& fcy,
+                       amrex::Array4<amrex::Real const> const& fcz),
+          amrex::GpuArray<amrex::Real,AMREX_SPACEDIM> const& dxinv)
+{
+    amrex::Real pdivu;
+
+#if (AMREX_SPACEDIM == 2)
+    if (flag(i,j,k).isRegular())
+    {
+       pdivu = 0.5 * ((q1(i + 1, j, k, GDPRES) + q1(i, j, k, GDPRES)) *
+                      (q1(i + 1, j, k, GDU)  - q1(i, j, k, GDU) ) *dxinv[0]
+                     +(q2(i, j + 1, k, GDPRES) + q2(i, j, k, GDPRES)) *
+                      (q2(i, j + 1, k, GDV)  - q2(i, j, k, GDV) ) * dxinv[1]);
+
+        divu(i,j,k,UEINT) += pdivu;
+    }
+    else
+    {
+        amrex::Real fxm = q1(i,j,k,GDU);
+        if (apx(i,j,k) != 0.0 && apx(i,j,k) != 1.0) {
+            int jj = j + static_cast<int>(amrex::Math::copysign(1.0, fcx(i,j,k,0)));
+            amrex::Real fracy = flag(i,j,k).isConnected(0,jj-j,0) ? amrex::Math::abs(fcx(i,j,k,0)) : 0.0;
+            fxm = (1.0-fracy)*fxm + fracy *q1(i,jj,k ,GDU);
+        }
+
+        amrex::Real fxp = q1(i+1,j,k,GDU);
+        if (apx(i+1,j,k) != 0.0 && apx(i+1,j,k) != 1.0) {
+            int jj = j + static_cast<int>(amrex::Math::copysign(1.0,fcx(i+1,j,k,0)));
+            amrex::Real fracy = flag(i+1,j,k).isConnected(0,jj-j,0) ? amrex::Math::abs(fcx(i+1,j,k,0)) : 0.0;
+            fxp = (1.0-fracy)*fxp + fracy *q1(i+1,jj,k,GDU);
+
+        }
+
+        amrex::Real fym = q2(i,j,k,GDV);
+        if (apy(i,j,k) != 0.0 && apy(i,j,k) != 1.0) {
+            int ii = i + static_cast<int>(amrex::Math::copysign(1.0,fcy(i,j,k,0)));
+            amrex::Real fracx = flag(i,j,k).isConnected(ii-i,0,0) ? amrex::Math::abs(fcy(i,j,k,0)) : 0.0;
+            fym = (1.0-fracx)*fym +  fracx *q2(ii,j,k,GDV);
+        }
+
+        amrex::Real fyp = q2(i,j+1,k,GDV);
+        if (apy(i,j+1,k) != 0.0 && apy(i,j+1,k) != 1.0) {
+            int ii = i + static_cast<int>(amrex::Math::copysign(1.0,fcy(i,j+1,k,0)));
+            amrex::Real fracx = flag(i,j+1,k).isConnected(ii-i,0,0) ? amrex::Math::abs(fcy(i,j+1,k,0)) : 0.0;
+            fyp = (1.0-fracx)*fyp + fracx *q2(ii,j+1,k,GDV);
+        }
+
+        divu(i,j,k,UEINT) += (1.0/vfrc(i,j,k)) * q(i,j,k,QPRES)*
+            ( dxinv[0] * (apx(i+1,j,k)*fxp-apx(i,j,k)*fxm)
+            + dxinv[1] * (apy(i,j+1,k)*fyp-apy(i,j,k)*fym) );
+
+    }
+
+#else // 3-d starts here
+
+    if (flag(i,j,k).isRegular())
+    {
+       pdivu = 0.5 * ((q1(i+1,j,k, GDPRES) + q1(i,j,k, GDPRES)) *
+                      (q1(i+1,j,k, GDU   ) - q1(i,j,k, GDU   )) *dxinv[0]
+                     +(q2(i,j+1,k, GDPRES) + q2(i,j,k, GDPRES)) *
+                      (q2(i,j+1,k, GDV   ) - q2(i,j,k, GDV   )) * dxinv[1]
+                     +(q2(i,j,k+1, GDPRES) + q2(i,j,k, GDPRES)) *
+                      (q2(i,j,k+1, GDW   ) - q2(i,j,k, GDW   )) * dxinv[2]);
+
+        divu(i,j,k,UEINT) += pdivu;
+    }
+    else
+    {
+        amrex::Real fxm = q1(i,j,k,GDU);
+        if (apx(i,j,k) != 0.0 && apx(i,j,k) != 1.0) {
+            int jj = j + static_cast<int>(amrex::Math::copysign(1.0, fcx(i,j,k,0)));
+            int kk = k + static_cast<int>(amrex::Math::copysign(1.0, fcx(i,j,k,1)));
+            amrex::Real fracy = flag(i,j,k).isConnected(0,jj-j,0) ? amrex::Math::abs(fcx(i,j,k,0)) : 0.0;
+            amrex::Real fracz = flag(i,j,k).isConnected(0,0,kk-k) ? amrex::Math::abs(fcx(i,j,k,1)) : 0.0;
+            fxm = (1.0-fracy)*(1.0-fracz)*fxm
+                +      fracy *(1.0-fracz)*q1(i,jj,k ,GDU)
+                +      fracz *(1.0-fracy)*q1(i,j ,kk,GDU)
+                +      fracy *     fracz *q1(i,jj,kk,GDU);
+        }
+
+        amrex::Real fxp = q1(i+1,j,k,GDU);
+        if (apx(i+1,j,k) != 0.0 && apx(i+1,j,k) != 1.0) {
+            int jj = j + static_cast<int>(amrex::Math::copysign(1.0,fcx(i+1,j,k,0)));
+            int kk = k + static_cast<int>(amrex::Math::copysign(1.0,fcx(i+1,j,k,1)));
+            amrex::Real fracy = flag(i+1,j,k).isConnected(0,jj-j,0) ? amrex::Math::abs(fcx(i+1,j,k,0)) : 0.0;
+            amrex::Real fracz = flag(i+1,j,k).isConnected(0,0,kk-k) ? amrex::Math::abs(fcx(i+1,j,k,1)) : 0.0;
+            fxp = (1.0-fracy)*(1.0-fracz)*fxp
+                +      fracy *(1.0-fracz)*q1(i+1,jj,k ,GDU)
+                +      fracz *(1.0-fracy)*q1(i+1,j ,kk,GDU)
+                +      fracy *     fracz *q1(i+1,jj,kk,GDU);
+
+        }
+
+        amrex::Real fym = q2(i,j,k,GDV);
+        if (apy(i,j,k) != 0.0 && apy(i,j,k) != 1.0) {
+            int ii = i + static_cast<int>(amrex::Math::copysign(1.0,fcy(i,j,k,0)));
+            int kk = k + static_cast<int>(amrex::Math::copysign(1.0,fcy(i,j,k,1)));
+            amrex::Real fracx = flag(i,j,k).isConnected(ii-i,0,0) ? amrex::Math::abs(fcy(i,j,k,0)) : 0.0;
+            amrex::Real fracz = flag(i,j,k).isConnected(0,0,kk-k) ? amrex::Math::abs(fcy(i,j,k,1)) : 0.0;
+            fym = (1.0-fracx)*(1.0-fracz)*fym
+                +      fracx *(1.0-fracz)*q2(ii,j,k ,GDV)
+                +      fracz *(1.0-fracx)*q2(i ,j,kk,GDV)
+                +      fracx *     fracz *q2(ii,j,kk,GDV);
+        }
+
+        amrex::Real fyp = q2(i,j+1,k,GDV);
+        if (apy(i,j+1,k) != 0.0 && apy(i,j+1,k) != 1.0) {
+            int ii = i + static_cast<int>(amrex::Math::copysign(1.0,fcy(i,j+1,k,0)));
+            int kk = k + static_cast<int>(amrex::Math::copysign(1.0,fcy(i,j+1,k,1)));
+            amrex::Real fracx = flag(i,j+1,k).isConnected(ii-i,0,0) ? amrex::Math::abs(fcy(i,j+1,k,0)) : 0.0;
+            amrex::Real fracz = flag(i,j+1,k).isConnected(0,0,kk-k) ? amrex::Math::abs(fcy(i,j+1,k,1)) : 0.0;
+            fyp = (1.0-fracx)*(1.0-fracz)*fyp
+                +      fracx *(1.0-fracz)*q2(ii,j+1,k ,GDV)
+                +      fracz *(1.0-fracx)*q2(i ,j+1,kk,GDV)
+                +      fracx *     fracz *q2(ii,j+1,kk,GDV);
+        }
+
+        amrex::Real fzm = q3(i,j,k,GDW);
+        if (apz(i,j,k) != 0.0 && apz(i,j,k) != 1.0) {
+            int ii = i + static_cast<int>(amrex::Math::copysign(1.0,fcz(i,j,k,0)));
+            int jj = j + static_cast<int>(amrex::Math::copysign(1.0,fcz(i,j,k,1)));
+            amrex::Real fracx = flag(i,j,k).isConnected(ii-i,0,0) ? amrex::Math::abs(fcz(i,j,k,0)) : 0.0;
+            amrex::Real fracy = flag(i,j,k).isConnected(0,jj-j,0) ? amrex::Math::abs(fcz(i,j,k,1)) : 0.0;
+
+            fzm = (1.0-fracx)*(1.0-fracy)*fzm
+                +      fracx *(1.0-fracy)*q3(ii,j ,k,GDW)
+                +      fracy *(1.0-fracx)*q3(i ,jj,k,GDW)
+                +      fracx *     fracy *q3(ii,jj,k,GDW);
+        }
+
+        amrex::Real fzp = q3(i,j,k+1,GDW);
+        if (apz(i,j,k+1) != 0.0 && apz(i,j,k+1) != 1.0) {
+            int ii = i + static_cast<int>(amrex::Math::copysign(1.0,fcz(i,j,k+1,0)));
+            int jj = j + static_cast<int>(amrex::Math::copysign(1.0,fcz(i,j,k+1,1)));
+            amrex::Real fracx = flag(i,j,k+1).isConnected(ii-i,0,0) ? amrex::Math::abs(fcz(i,j,k+1,0)) : 0.0;
+            amrex::Real fracy = flag(i,j,k+1).isConnected(0,jj-j,0) ? amrex::Math::abs(fcz(i,j,k+1,1)) : 0.0;
+            fzp = (1.0-fracx)*(1.0-fracy)*fzp
+                +      fracx *(1.0-fracy)*q3(ii,j ,k+1,GDW)
+                +      fracy *(1.0-fracx)*q3(i ,jj,k+1,GDW)
+                +      fracx *     fracy *q3(ii,jj,k+1,GDW);
+        }
+
+        divu(i,j,k,UEINT) += (1.0/vfrc(i,j,k)) * q(i,j,k,QPRES)*
+            ( dxinv[0] * (apx(i+1,j,k)*fxp-apx(i,j,k)*fxm)
+            + dxinv[1] * (apy(i,j+1,k)*fyp-apy(i,j,k)*fym)
+            + dxinv[2] * (apz(i,j,k+1)*fzp-apz(i,j,k)*fzm) );
+    }
+#endif
+}
 
 AMREX_GPU_DEVICE AMREX_FORCE_INLINE
 void eb_compute_div (int i, int j, int k, int n,
@@ -295,3 +533,5 @@ void eb_compute_div (int i, int j, int k, int n,
     else if (eb_weights_type == 3)
         { redistwgt(i,j,k) = vfrc(i,j,k); }
 }
+#endif
+#endif
