+++ Source/Hydro/Godunov/Godunov_2D.cpp
@@ -1,11 +1,13 @@
 #include "Godunov.H"
+#include "CAMR_utils_K.H"
+#include "Godunov_utils.H"
 #include "Hydro_cmpflx.H"
 #include "PLM.H"
 #include "PPM.H"
 
 #if (AMREX_SPACEDIM == 2)
 void
-CAMR_umeth_2D(
+Godunov_umeth (
   amrex::Box const& bx,
   const int* bclo,
   const int* bchi,
@@ -33,7 +35,7 @@ CAMR_umeth_2D(
   const int iorder,
   const int l_transverse_reset_density)
 {
-  BL_PROFILE("CAMR::CAMR_umeth_2D()");
+  BL_PROFILE("CAMR::Godunov_umeth()");
   amrex::Real const dx = del[0];
   amrex::Real const dy = del[1];
   amrex::Real const hdt = 0.5 * dt;
