+++ Source/Hydro/Godunov/Make.package
@@ -2,6 +2,7 @@
 CEXE_sources += Godunov_$(DIM)D.cpp
 CEXE_sources += PPM.cpp
 CEXE_headers += Godunov.H
+CEXE_headers += Godunov_utils.H
 CEXE_headers += PLM.H
 CEXE_headers += PPM.H
